Extract get args
