+++ src/actions/get-nonce.ts
@@ -1,12 +1,16 @@
 import { makeApplication } from "../boot.js";
 import { loadConfig } from "../loader.js";
 import { makeIdentityFactories } from "../builder.js";
+import { getArgs } from "../utils.js";
 
 import * as Client from "../client.js";
 
-const main = async () => {
+const main = async (customArgs?: string[]) => {
     const config = loadConfig();
-    const passphrase = process.argv.length === 3 ? process.argv[2] : config.cli.senderPassphrase;
+
+    const { args } = getArgs(customArgs);
+
+    const passphrase = args.length === 1 ? args[0] : config.cli.senderPassphrase;
     const peer = config.cli.peer;
 
     const app = await makeApplication(config);
